Refactor Angular app with external templates
