+++ ui/src/app/app.component.css
@@ -1,0 +1,9 @@
+table {
+  border-collapse: collapse;
+  margin: 0 auto;
+}
+
+th, td {
+  border: 1px solid #ccc;
+  padding: 4px 8px;
+}
