+++ ui/src/app/app.component.html
@@ -1,0 +1,22 @@
+<h1>Health Check UI</h1>
+<h2>URLs</h2>
+<ul>
+  <li *ngFor="let u of urls">{{u.target}}</li>
+</ul>
+<h2>Records</h2>
+<table>
+  <thead>
+    <tr>
+      <th>URL</th>
+      <th>Status</th>
+      <th>Timestamp</th>
+    </tr>
+  </thead>
+  <tbody>
+    <tr *ngFor="let r of records">
+      <td>{{r.url?.target}}</td>
+      <td>{{r.status_code}}</td>
+      <td>{{r.timestamp}}</td>
+    </tr>
+  </tbody>
+</table>
